Rename transform_map module
